# api/api/openapi.yml
openapi: '3.0.3'
info:
  title: AusCERT 2023 - Going back to the basics of API security
  description: API documentation for a demo API.
  contact:
    name: Kyle Jackson
    email: kyle.jackson@forgeresponse.com
  version: 1.0.0
servers:
  - url: https://auscert.forgeresponse.com
components:
  schemas:
...
    Incident:
      type: object
      properties:
        id:
          type: string
        assignedTo:
          type: string
        attachments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
        createdBy:
          type: string
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        tasks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
        title:
          type: string
        tlp:
          type: integer
    CreateIncidentRequest:
      type: object
      properties:
        assignedTo:
          type: string
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        tlp:
          type: integer
    UpdateIncidentRequest:
      type: object
      properties:
        assignedTo:
          type: string
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        tlp:
          type: integer
    IncidentComment:
      type: object
      properties:
        id:
          type: string
        comment:
          type: string
        createdAt:
          type: string
        createdBy:
          type: string
        order:
          type: integer
    CreateIncidentCommentRequest:
      type: object
      properties:
        comment:
          type: string
    UpdateIncidentCommentRequest:
      type: object
      properties:
        comment:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        code:
          type: integer
        transactionId:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header
security:
  - ApiKeyAuth: []
paths:
...
  /v1/incidents:
    get:
      description: Returns a list of incidents
      tags:
        - Incidents
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Specify the number of items to return
      responses:
        '200':
          description: Successfully returned a list of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incident'
        '404':
          description: No records found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      description: Creates a new incident
      tags:
        - Incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
      responses:
        '200':
          description: Successfully created incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{id}:
    parameters:
      - name: id
        in: path
        description: Id of the incident
        required: true
        schema:
          type: string
    get:
      description: Returns the specified incident
      tags:
        - Incidents
      responses:
        '200':
          description: Successfully returned incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      description: Updates the specified incident
      tags:
        - Incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentRequest'
      responses:
        '200':
          description: Successfully updated incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      description: Deletes the specified incident
      tags:
        - Incidents
      responses:
        '204':
          description: Successfully deleted incident
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{id}/comments:
    parameters:
      - name: id
        in: path
        description: Id of the incident
        required: true
        schema:
          type: string
    get:
      description: Returns a list of incident comments
      tags:
        - Incidents
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Specify the number of items to return
      responses:
        '200':
          description: Successfully returned a list of incident comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncidentComment'
        '404':
          description: No records found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      description: Creates a new incident comment
      tags:
        - Incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentCommentRequest'
      responses:
        '200':
          description: Successfully created incident comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentComment'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{id}/comments/{commentId}:
    parameters:
      - name: id
        in: path
        description: Id of the incident
        required: true
        schema:
          type: string
      - name: commentId
        in: path
        description: Id of the comment
        required: true
        schema:
          type: string
    get:
      description: Returns the specified incident comment
      tags:
        - Incidents
      responses:
        '200':
          description: Successfully returned incident comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentComment'
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      description: Updates the specified incident comment
      tags:
        - Incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentCommentRequest'
      responses:
        '200':
          description: Successfully updated incident comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentComment'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      description: Deletes the specified incident comment
      tags:
        - Incidents
      responses:
        '204':
          description: Successfully deleted incident comment
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...